refactor(cli): thin telegram.ts — sort logic to core/telegram, keep stdin at edge (S2, #367)
Extract 14 functions/1 interface from cli/commands/telegram.ts to core:
- core/telegram/notify.ts (new): TelegramStatusResult, runTelegramStatus,
  runTelegramNotifySet, runTelegramNotifyPref, runTelegramNotifyStatus,
  capAllowed, runNotifyConfirmation, runTelegramSend, runTelegramLink
- core/telegram/setup.ts (extended): resolveSetupToken, resolveSetupUserId,
  runRegisterCommands, runTelegramPostSetup

telegram.ts shrinks from 565 → 307 LOC. Re-exports keep existing CLI tests green.
